critical boot time fix
Bot would constantly go offline due to the fact that the VerifyServer database operation was extremely poorly optimized. So much in fact... that say we have a server with 300 members. For each member in that server, the bot would iterate through every member 300^4 (81 million) times... yeah, not good.... and that's assuming my math is correct. The new framework only iterates through all members once, so boot times are SIGNIFICANTLY reduced. Furthermore, the bot verifies the server database after many common operations, such as role additions, so the fact that long downtimes existed due to this were problematic. I apologize for the previous inconveniences caused - the bot will remain ready to serve you now