patch 0.2.24.7
•Added a dev-only command to remove old index data
    •Removes all data entirely for servers the bot is no longer in
    •Removes the thousands of text files left over from the old indexes system. This is something that is essential to do, but I implemented it now because I plan on using my new Raspberry Pi 4 to host Disguard - and transferring the local data will be extremely impractical with thousands of text files in the system (one of many advantages of the .json file system)

These processes will be taken care of automatically in the next major update, along with the bot also deleting database entries for old data. Lack of removal of index data was just  never implemented, whereas lack of removal of database data seems to be due to an error, as it's supposed to happen.

Once I clean up the logging module and finish the new features, I'll be working on improvements to the website/dashboard, but primarily a privacy system as well (privacy policy, privacy settings/info, etc) before releasing Disguard on top.gg and other listing sites. Note that work on modules other than logging (such as moderation commands and the antispam system) and a work for the future - for now, Disguard is still (marginally) better than Ricobot, at worst, and has new features/improvements in many other places.

Last update of 2020! Farewell 2020! 2021 will being a lot for Disguard due to increased programming motivation and a fresh new year/college semester in February for me.