patch 0.2.28.1
•Removed memory tracker diagnostic information

For reference, the Disguard bot process takes up around 1.3 to 1.5GB of RAM. Most of it is used by the constant database operations, along with caching the entire database in RAM.

The first one is caused by unoptimized and outdated code from me in 2019 in Disguard's early days. I'll make some changes over the next few days, but only to ensure Disguard doesn't "hang" as much. The rest will go live with Disguard v1.0.

As for the database caching, I'll probably have to move it to local storage or a local database or something. We'll see how much time I have, but I'm prioritizing performance before eye candy for the launch of Disguard v1.0.