patch 0.2.30
•Fix critical security flaw

Rundown: Due to incorrect class converting with arguments passed to functions, a number of users (as high as 46 by the time this fix goes out) were able to manage any single server Disguard was in from the online dashboard. Due to differences in how a discord Member (belongs to a server) and discord User (doesn't belong to a server) behave in discord.py, my code was functioning incorrectly when a discord User was used to determine which servers people were allowed to manage. This bug has been rectified in this update.

Big thanks to PVZABFan for reporting this bug to me. As soon as this bug was reported, I locked web dashboard access to all users, since the number of users able to manage all servers was growing every so often. Now that this update is going out, I will unlock the web dashboard again.