patch 0.2.30
A while ago, it was suggested to me that I could make use of the Discord CDN storage for attachments < 8mb instead of storing them all on the local filesystem. Due to running out of space on the device running Disguard, I have implemented this system to store new attachments. If I add more storage to the host device, or if I regain enough storage space back, I may consider partially using the local filesystem again.

-----New features/Changes-----
•Added new system for storing attachments
    •Instead of writing attachments to the host machine's local filesystem, attachments are saved to a private channel, and a directory linking to those messages is maintained to retrieve attachments upon message deletion
    •The method for saving attachments (when enabled) has been replaced by the new method, but the method for retrieving attachments searches the local filesystem before searching the new cache channel
-----Minor changes/Bugfixes-----
•Changed message index keys before saving to the json file to strings to prevent potential mishaps
•Changed behavior of message indexing loop on bootup to only move on if 100 messages in the current channel in a row are already in the index file