improve UX for setting age/birthday
if a user tries to update their age/birthday to the value already set, the program will now inform the user that their age/birthday is already set to this value. previously, the program would either return (meaning on the UX side, the user would get ghosted) and would only inform the user if absolutely necessary.