refactor data storage to support local MongoDB
remove the "lightningLogging" and "lightningUsers" cache data that have been in use for 2-3 years to favor a local MongoDB database solution. the cache feature previously used too much memory causing disguard to be unstable in the long run. there may be a *very slight* performance hit with the database implementation, but it's a much better solution to the data storage.

this is purely a backend change and probably won't be noticed at all by the end users.

the next step is for me to replace the current json implementation of message indexing with the database version. that change should bring performance improvements, improved flexibility, and many more benefits.

after that point, i'll start preparing these development commits for the next major public update for the bot.

also some other minor changes/fixes:
•change generator type with data sync method to prevent unnecessary users from appearing in the bot's members list (such as those who left a server)