patch 0.5.8
•Rewrite the database method that processes server methods to simplify the code, reduce memory usage, and reduce operation time
•Fix an issue preventing calculateGeneralChannel from working
•The bot will update its status on bootup to say "booted" after it finishes indexing messages rather than hanging on "indexing..." while fetching channel pins (which is a very long operation due to API ratelimits)
•Make other improvements to the CalculateModeratorChannel and CalculateGeneralChannel functions to reduce memory usage by storing ints rather than discord channel objects
•Improve print statement diagnostics for the update server database function

this is kind of a band-aid fix in some aspects, a much better fix will be for me to change the data type of server members from a list to a dict. this will need a lot more infrastructure updated to reflect that major change, which will likely take a couple of days to finish at least. until then, this update has some very notable performance improvements to hopefully keep the bot online longer before crashing.